Complete HW4 JWT authentication

+++ src/controllers/announcements.controller.js
@@ -10,7 +10,7 @@ export const getAnnouncements = async (req, res) => {
 
   if (search) {
     where.title = {
-    contains: search,
+      contains: search,
     }
   }
 
@@ -53,7 +53,10 @@ export const getAnnouncementById = async (req, res) => {
 export const createAnnouncement = async (req, res) => {
   const announcement =
     await prisma.announcement.create({
-      data: req.body,
+      data: {
+        ...req.body,
+        userId: req.user.id,
+      },
     })
 
   res.status(201).json(announcement)
@@ -61,6 +64,19 @@ export const createAnnouncement = async (req, res) => {
 
 export const updateAnnouncement = async (req, res) => {
   const announcement =
+    await prisma.announcement.findUniqueOrThrow({
+      where: {
+        id: Number(req.params.id),
+      },
+    })
+
+  if (announcement.userId !== req.user.id) {
+    return res.status(403).json({
+      error: 'Access denied',
+    })
+  }
+
+  const updatedAnnouncement =
     await prisma.announcement.update({
       where: {
         id: Number(req.params.id),
@@ -68,10 +84,23 @@ export const updateAnnouncement = async (req, res) => {
       data: req.body,
     })
 
-  res.json(announcement)
+  res.json(updatedAnnouncement)
 }
 
 export const deleteAnnouncement = async (req, res) => {
+  const announcement =
+    await prisma.announcement.findUniqueOrThrow({
+      where: {
+        id: Number(req.params.id),
+      },
+    })
+
+  if (announcement.userId !== req.user.id) {
+    return res.status(403).json({
+      error: 'Access denied',
+    })
+  }
+
   await prisma.announcement.delete({
     where: {
       id: Number(req.params.id),
